product manage: add and query completed 1. will add instore outstore 2. category show as tree,using ztree 3. 调整拦截器，session不再存储permission，后续添加redis 和spring decurity 4. 出库单刚开始做

+++ src/main/java/rk/service/ProductService.java
@@ -4,6 +4,7 @@ import org.json.JSONObject;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import rk.base.BaseService;
+import rk.configuration.enuma.OperationStatus;
 import rk.dao.ProductDao;
 import rk.exceptions.ParamRequestException;
 import rk.model.ResultInfo;
@@ -34,11 +35,7 @@ public class ProductService extends BaseService<Product> {
         ArrayList<Integer> ids = new ArrayList<>();
         ids.add(productCategory.getId());
         for(ProductCategory childProductCategory:productCategory.getChildren()){
-            if(childProductCategory.getChildren().size()>0){
-                ids.addAll( setChildIds( childProductCategory ) );
-            }else{
-                ids.add(childProductCategory.getId());
-            }
+            ids.addAll( setChildIds( childProductCategory ) );
         }
         productCategory.setChildIds( ids );
         return ids;
@@ -49,16 +46,21 @@ public class ProductService extends BaseService<Product> {
     }
     @Transactional
     public void insertOrUpdateProduct(Product product) {
+
+        if(product.getProductId() != null) {
+            AssertUtil.isTrue( productDao.deleteProductSpecifications( product.getProductId() ) < 1, "产品规格删除失败" );
+        }
+
         AssertUtil.isTrue( this.saveUpdate( product,product.getProductId() )<1,"产品添加失败" );
         for(int i=0;i < product.getProductSpecifications().size();i++){
             ProductSpecification productSpecification = product.getProductSpecifications().get( i );
-            String name = productSpecification.getSpecificationName();
-            if("".equals( name )||null == name){
-                throw new ParamRequestException( 300, "产品规格名不能为空" );
+            if("".equals( productSpecification.getSpecificationName() )||null == productSpecification.getSpecificationName()){
+                throw new ParamRequestException( OperationStatus.paramNotAvailable );
+            }else if("".equals( productSpecification.getPrice() )||null == productSpecification.getPrice()){
+                throw new ParamRequestException( OperationStatus.paramNotAvailable );
             }
             productSpecification.setProductId( product.getProductId() );
         }
-        AssertUtil.isTrue( productDao.deleteProductSpecifications( product.getProductId() )<1,"产品规格删除失败" );
         AssertUtil.isTrue( productDao.insertproductSpecifications(product.getProductSpecifications())<1,"产品规格添加失败" );
     }
 
@@ -66,5 +68,9 @@ public class ProductService extends BaseService<Product> {
     public void deleteProduct(Integer productId) {
         this.delete( productId );
         AssertUtil.isTrue( productDao.deleteProductSpecifications( productId )<1,"删除失败" );
+    }
+
+    public List<ProductSpecification> queryProductSpecifications(Integer productId) {
+        return productDao.queryProductSpecificationByProductId( productId );
     }
 }
